pyproject.toml
[project]
name = "back-to-basics"
version = "1.0.0"
description = "just releaning fastapi basics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.118.0",
    "logfire>=4.11.0",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.11.10",
    "python-dotenv>=1.1.1",
    "supabase>=2.21.1",
    "uvicorn>=0.37.0",
]
